ci: run integration tests with code changes before releasing new action
This GH Action workflow runs integration tests on `push` and `pull_request` with the latest
code of this `rust-action`. It does require neither a release nor a push to master to test
the latest changes in this action. This workflow can be used as required test for a branch
protection rule.